HomeViewをStackに変更、一部コンポーネント修正
