+++ cafelogos-pos/Presentation/HomeView/Components/TrainingStatus.swift
@@ -1,0 +1,18 @@
+//
+//  TrainingStatus.swift
+//  cafelogos-pos
+//
+//  Created by Owner on 2023/09/29.
+//
+
+import SwiftUI
+
+struct TrainingStatus: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TrainingStatus()
+}
