+++ cafelogos-pos/Presentation/HomeView/HomeView.swift
@@ -11,21 +11,6 @@ struct HomeView: View {
     @State private var displayConnection: Bool = true // true: 接続中, false: 切断中
     @State private var serverConnection: Bool = true // true: 接続中, false: 切断中
     
-    let homeDestinations: [HomeDestination<AnyView>] = [
-        HomeDestination(title: "点検・精算", subTitle: "", description: "", destination: {AnyView(SettingView())}, fg_color: Color.primary, bg_color: Color(.systemFill)),
-        HomeDestination(title: "レジ開け", subTitle: "", description: "", destination: {AnyView(StartTransactionView())}, fg_color: Color.primary, bg_color: Color(.systemFill)),
-        HomeDestination(title: "設定", subTitle: "", description: "", destination: {AnyView(SettingView())}, fg_color: Color.primary, bg_color: Color(.systemFill)),
-        HomeDestination(title: "トレーニング", subTitle: "オン・オフ切り替え", description: "", destination: {AnyView(SettingView())}, fg_color: Color.primary, bg_color: Color(.systemFill)),
-        HomeDestination(title: "トレーニング", subTitle: "オン・オフ切り替え", description: "", destination: {AnyView(SettingView())}, fg_color: Color.primary, bg_color: Color(.systemFill)),
-    ]
-    let gridItems = [
-        GridItem(.adaptive(minimum: 100, maximum: 10000),spacing: 15,alignment: .top ),
-        GridItem(.flexible(),spacing: 15,alignment: .top),
-        GridItem(.flexible(),spacing: 15,alignment: .top),
-        GridItem(.flexible(),spacing: 15,alignment: .top),
-        //        GridItem(.flexible(),spacing: 15),
-    ]
-    
     
     
     
@@ -48,30 +33,12 @@ struct HomeView: View {
                         
                         // 右列
                         VStack(spacing: 15){
-                            HomeNavButton(title: "点検", subtitle: "", description: "", destination: {OrderInputView(productQueryService: ProductQueryServiceMock(), discountRepository: DiscountRepositoryMock())}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
-                            HomeNavButton(title: "精算", subtitle: "", description: "", destination: {OrderInputView(productQueryService: ProductQueryServiceMock(), discountRepository: DiscountRepositoryMock())}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
-                            HomeNavButton(title: "設定", subtitle: "", description: "", destination: {OrderInputView(productQueryService: ProductQueryServiceMock(), discountRepository: DiscountRepositoryMock())}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
+                            HomeNavButton(title: "点検", subtitle: "", description: "", destination: {SettlementView()}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
+                            HomeNavButton(title: "精算", subtitle: "", description: "", destination: {SettlementView()}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
+                            HomeNavButton(title: "設定", subtitle: "", description: "", destination: {SettingView()}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
                             HomeNavButton(title: "トレーニング", subtitle: "", description: "オン・オフ切り替え", destination: {OrderInputView(productQueryService: ProductQueryServiceMock(), discountRepository: DiscountRepositoryMock())}, fg_color: Color.primary, bg_color: Color(.systemFill), height: geometry.size.height * (1/4) , width: geometry.size.width * (1/3))
                         }
-                    
-                        LazyHGrid(rows: gridItems, spacing: 15 ) {
-                            ForEach(homeDestinations.indices, id: \.self) { index in
-                                HomeNavButton(
-                                    title: homeDestinations[index].title,
-                                    subtitle: homeDestinations[index].subTitle,
-                                    description: homeDestinations[index].description,
-                                    destination: {homeDestinations[index].destination()},
-                                    fg_color: homeDestinations[index].fg_color,
-                                    bg_color: homeDestinations[index].bg_color,
-                                    height: geometry.size.height * (1/4),
-                                    width: geometry.size.width * (1/3)
-                                    
-                                )
-                                
-                            }
-                        }
-                        
-                        .frame(maxWidth: .infinity, maxHeight: .infinity)
+
                     }
                     .padding(20)
                     .frame(maxWidth: .infinity, maxHeight: .infinity)
@@ -81,6 +48,7 @@ struct HomeView: View {
             }
             .frame(maxWidth: .infinity, maxHeight: .infinity)
         }
+        .navigationBarBackButtonHidden(true)
     }
 }
