+++ cafelogos-pos/Presentation/Transactions/StartTransactionView.swift
@@ -25,7 +25,7 @@ struct StartTransactionView: View {
                             .padding(.top, 50)
                         Spacer()
                         multi2Button(title: "スキップする", subtitle: "(トレーニングモード)", bgColor: Color.secondary, fgColor: Color.white, destination: {HomeView()})
-                        multi1Button(title: "次へ", bgColor: Color.cyan, fgColor: Color.white, destination: {HomeView()})
+                        TitleButton(title: "次へ", bgColor: Color.cyan, fgColor: Color.white, destination: {HomeView()})
                             
                             .padding(.bottom)
                        
@@ -39,42 +39,43 @@ struct StartTransactionView: View {
                 }
                 .frame(maxWidth: .infinity, maxHeight: .infinity)
             }
+            
         }
     }
 }
-struct multi1Button<Destination>: View where Destination : View {
-    var title: String
-    var bgColor: Color
-    var fgColor: Color
-    var destination: () -> Destination
-    
-    public init(title: String, bgColor: Color, fgColor: Color, destination: @escaping () -> Destination) {
-        self.title = title
-        self.bgColor = bgColor
-        self.fgColor = fgColor
-        self.destination = destination
-    }
-    
-    var body: some View {
-        NavigationLink(destination: destination) {
-            VStack(spacing: 0) {
-                Text(title)
-                    .font(.largeTitle)
-                    .fontWeight(.semibold)
-                    .lineLimit(0)
-            }
-            .foregroundColor(fgColor)
-            .frame(maxWidth: .infinity, minHeight: 50, alignment: .center)
-            .clipped()
-            .padding(.vertical, 20)
-            .background {
-                RoundedRectangle(cornerRadius: 8, style: .continuous)
-                    .fill(bgColor)
-            }
-            .padding(.bottom)
-        }
-    }
-}
+//struct multi1Button<Destination>: View where Destination : View {
+//    var title: String
+//    var bgColor: Color
+//    var fgColor: Color
+//    var destination: () -> Destination
+//    
+//    public init(title: String, bgColor: Color, fgColor: Color, destination: @escaping () -> Destination) {
+//        self.title = title
+//        self.bgColor = bgColor
+//        self.fgColor = fgColor
+//        self.destination = destination
+//    }
+//    
+//    var body: some View {
+//        NavigationLink(destination: destination) {
+//            VStack(spacing: 0) {
+//                Text(title)
+//                    .font(.largeTitle)
+//                    .fontWeight(.semibold)
+//                    .lineLimit(0)
+//            }
+//            .foregroundColor(fgColor)
+//            .frame(maxWidth: .infinity, minHeight: 50, alignment: .center)
+//            .clipped()
+//            .padding(.vertical, 20)
+//            .background {
+//                RoundedRectangle(cornerRadius: 8, style: .continuous)
+//                    .fill(bgColor)
+//            }
+//            .padding(.bottom)
+//        }
+//    }
+//}
 
 struct multi2Button<Destination>: View where Destination : View {
     var title: String
